+++ tomcat/src/main/java/org/apache/catalina/ExceptionHandler.java
@@ -1,0 +1,8 @@
+package org.apache.catalina;
+
+import org.apache.coyote.controller.Controller;
+
+public interface ExceptionHandler {
+
+    Controller getHandler(final Exception exception);
+}
